ai_ml/requirements.txt
# AI/ML Requirements for Hospital Supply Chain Platform

# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0

# LLM Integration - Gemini API
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Data visualization (optional but useful)
matplotlib>=3.4.0
seaborn>=0.11.0

# Additional utilities
typing-extensions>=4.0.0

# Performance monitoring and caching
redis>=4.5.0  # Optional: for distributed caching
aiofiles>=23.0.0  # For async file operations

# Note: For production deployment, consider adding:
# tensorflow>=2.8.0  # For deep learning models
# torch>=1.11.0      # Alternative deep learning framework
# statsmodels>=0.13.0 # For advanced time series analysis
# plotly>=5.0.0       # Interactive visualizations
